fix: use usize for pruned counters to avoid unnecessary cast
Amp-Thread-ID: https://ampcode.com/threads/T-019c115f-df26-753b-9b4d-32501c038860

+++ crates/cli/commands/src/prune.rs
@@ -79,11 +79,11 @@ impl<C: ChainSpecParser<ChainSpec: EthChainSpec + EthereumHardforks>> PruneComma
                     .delete_limit(DELETE_LIMIT)
                     .build_with_provider_factory(provider_factory);
 
-                let mut total_pruned = 0u64;
+                let mut total_pruned = 0usize;
                 loop {
                     let output = pruner.run(prune_tip)?;
-                    let batch_pruned: u64 =
-                        output.segments.iter().map(|(_, seg)| seg.pruned as u64).sum();
+                    let batch_pruned: usize =
+                        output.segments.iter().map(|(_, seg)| seg.pruned).sum();
                     total_pruned = total_pruned.saturating_add(batch_pruned);
 
                     if output.progress.is_finished() {
